requirements.txt
opencv-python
pillow
numpy
torch
facenet_pytorch